Moved BooleanToDecimal and DecimalToBoolean into OracleTools.
