+++ Tewl/Tools/DecimalTools.cs
@@ -66,16 +66,5 @@ namespace Tewl.Tools {
 			var centValue = dollarValue * 100;
 			return centValue != (int)centValue;
 		}
-
-		/// <summary>
-		/// Converts a decimal (presumably from Oracle) into a boolean.
-		/// </summary>
-		public static bool DecimalToBoolean( this decimal d ) {
-			if( d == 1 )
-				return true;
-			if( d == 0 )
-				return false;
-			throw new ApplicationException( "Unknown decimal value encountered when converting to boolean." );
-		}
 	}
 }
